FIXED Username checks + Locked Account PU
Started to create a limit on login attempts. Made a class in the popup file to lock the account for 30 seconds. Began to set to the limit to three. Found a issue where username was always returned as false. It wouldnt register the stored username and entered as the same as they where two different data types. Used a for loop to convert stored data to a string afetr it had been fetched. Fixed the Username check.